BUG/requirements.txt
Flask==3.1.0
Werkzeug>=3.1
Flask-Cors==3.0.10
gunicorn==20.1.0
tensorflow==2.18.0  # TensorFlow includes Keras
beautifulsoup4==4.11.1
requests==2.28.1
python-gitlab==3.10.0
nltk==3.7
pandas==1.5.3
numpy==1.26.0  # Updated numpy version
scikit-learn==1.1.2
